% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegJulich-package.R
\docType{package}
\name{ggsegJulich-package}
\alias{ggsegJulich}
\alias{ggsegJulich-package}
\title{ggsegJulich: Julich-Brain Cytoarchitectonic Atlas for the 'ggseg' Ecosystem}
\description{
Julich-Brain cytoarchitectonic maps of the human brain for the 'ggseg' ecosystem. Provides a unified 'ggseg_atlas' object with both 2D polygon geometry and 3D vertex indices, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegJulich}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegJulich/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
