% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{julich}
\alias{julich}
\title{Julich-Brain Cytoarchitectonic Atlas}
\usage{
julich()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Maximum probability map from 296 cytoarchitectonic probability maps
with 294 labelled regions.
}
\examples{
julich()
plot(julich())
}
\references{
Amunts K, et al. (2020). Julich-Brain: A 3D probabilistic
atlas of the human brain's cytoarchitecture. \emph{Science}, 369(6506):988-992.
\doi{10.1126/science.abb4588}
}
\concept{ggseg_atlases}
