% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{julich_cortical}
\alias{julich_cortical}
\title{Julich-Brain Cortical Atlas}
\usage{
julich_cortical()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Cortical regions from the maximum probability map of 296
cytoarchitectonic probability maps.
}
\examples{
julich_cortical()
}
\references{
Amunts K, et al. (2020). Julich-Brain: A 3D probabilistic
atlas of the human brain's cytoarchitecture. \emph{Science}, 369(6506):988-992.
\doi{10.1126/science.abb4588}
}
\seealso{
Other ggseg_atlases: 
\code{\link{julich_subcortical}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
