% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{julich_subcortical}
\alias{julich_subcortical}
\title{Julich-Brain Subcortical Atlas}
\usage{
julich_subcortical()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Subcortical regions from the maximum probability map of 296
cytoarchitectonic probability maps.
}
\examples{
julich_subcortical()
}
\references{
Amunts K, et al. (2020). Julich-Brain: A 3D probabilistic
atlas of the human brain's cytoarchitecture. \emph{Science}, 369(6506):988-992.
\doi{10.1126/science.abb4588}
}
\seealso{
Other ggseg_atlases: 
\code{\link{julich_cortical}()}
}
\concept{ggseg_atlases}
\concept{subcortical_atlases}
